tools/requirements_sonos.txt
soco>=0.30.0
pillow>=10.0.0
requests>=2.31.0 